run flask in the background
