+++ src/keygen.h
@@ -26,6 +26,4 @@ private:
     std::string key_format;
 };
 
-void uploadFileToCodespace(const std::string& filename, const std::string& codespaceName);
-
 #endif /* KEYGEN_H */
